+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/storage/disk/StorageDiskFluidPortable.java
@@ -6,18 +6,18 @@ import com.raoulvdberge.refinedstorage.api.storage.disk.IStorageDiskContainerCon
 import com.raoulvdberge.refinedstorage.api.storage.disk.IStorageDiskListener;
 import com.raoulvdberge.refinedstorage.api.util.Action;
 import com.raoulvdberge.refinedstorage.tile.grid.portable.IPortableGrid;
-import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
+import net.minecraftforge.fluids.FluidStack;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.Collection;
 
-public class StorageDiskItemPortable implements IStorageDisk<ItemStack> {
-    private IStorageDisk<ItemStack> parent;
+public class StorageDiskFluidPortable implements IStorageDisk<FluidStack> {
+    private IStorageDisk<FluidStack> parent;
     private IPortableGrid portableGrid;
 
-    public StorageDiskItemPortable(IStorageDisk<ItemStack> parent, IPortableGrid portableGrid) {
+    public StorageDiskFluidPortable(IStorageDisk<FluidStack> parent, IPortableGrid portableGrid) {
         this.parent = parent;
         this.portableGrid = portableGrid;
     }
@@ -38,22 +38,22 @@ public class StorageDiskItemPortable implements IStorageDisk<ItemStack> {
     }
 
     @Override
-    public Collection<ItemStack> getStacks() {
+    public Collection<FluidStack> getStacks() {
         return parent.getStacks();
     }
 
     @Nullable
     @Override
-    public ItemStack insert(@Nonnull ItemStack stack, int size, Action action) {
+    public FluidStack insert(@Nonnull FluidStack stack, int size, Action action) {
         int storedPre = parent.getStored();
 
-        ItemStack remainder = parent.insert(stack, size, action);
+        FluidStack remainder = parent.insert(stack, size, action);
 
         if (action == Action.PERFORM) {
             int inserted = parent.getCacheDelta(storedPre, size, remainder);
 
             if (inserted > 0) {
-                portableGrid.getItemCache().add(stack, inserted, false, false);
+                portableGrid.getFluidCache().add(stack, inserted, false, false);
             }
         }
 
@@ -62,11 +62,11 @@ public class StorageDiskItemPortable implements IStorageDisk<ItemStack> {
 
     @Nullable
     @Override
-    public ItemStack extract(@Nonnull ItemStack stack, int size, int flags, Action action) {
-        ItemStack extracted = parent.extract(stack, size, flags, action);
+    public FluidStack extract(@Nonnull FluidStack stack, int size, int flags, Action action) {
+        FluidStack extracted = parent.extract(stack, size, flags, action);
 
         if (action == Action.PERFORM && extracted != null) {
-            portableGrid.getItemCache().remove(extracted, extracted.getCount(), false);
+            portableGrid.getFluidCache().remove(extracted, extracted.amount, false);
         }
 
         return extracted;
@@ -88,7 +88,7 @@ public class StorageDiskItemPortable implements IStorageDisk<ItemStack> {
     }
 
     @Override
-    public int getCacheDelta(int storedPreInsertion, int size, @Nullable ItemStack remainder) {
+    public int getCacheDelta(int storedPreInsertion, int size, @Nullable FluidStack remainder) {
         return parent.getCacheDelta(storedPreInsertion, size, remainder);
     }